feat: 체스방 목록 조회하기 기능 구현
