+++ README.md
@@ -12,7 +12,7 @@
   - 체스방 만들기 버튼을 누르고 체스방 제목과 비밀번호를 입력하면 새로운 체스판이 만들어진다.
   - 체스방에는 고유식별값이 부여된다. (이 고유 식별값은 체스방 주소에서 사용 됨)
 
-- [ ] 체스방 목록 조회하기 
+- [x] 체스방 목록 조회하기 
   - localhost:8080 요청 시 체스방 목록을 조회할 수 있다
   - 체스방 목록에는 체스방 제목과 체스방을 삭제할 수 있는 버튼이 표시된다.
