+++ src/main/resources/templates/finished.hbs
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>체스 게임</title>
     <link rel="stylesheet" href="/src/index.css">
-    <script src="/src/index.js"></script>
+    <script src="/src/game.js"></script>
 </head>
 
 <body>
