+++ src/main/resources/templates/index.hbs
@@ -29,14 +29,16 @@
         <td><a href=""></a></td>
       </tr>
     </template>
+
     <table class="chess-room-list">
       <thead>
       <tr>
         <td>번호</td>
         <td>제목</td>
+        <td>삭제하기</td>
       </tr>
       </thead>
-      <tbody>
+      <tbody class="tbody-list">
       </tbody>
     </table>
   </section>
